[BKR-18] [BKR-13] [BKR-16] Not allowed to deposit funds directly to vault, Chainlink - Deprecated Integration And Lax Validation, Calculations On Stale Price Data During Withdrawal (#22)
* Test fixes

* BakerFiVault  with outdated prices

* BakerFi Vault tests with outdated prices

* Prettier

* Calculations on stale prices

* Working with custom errors

* Strategy

* Linter run

* Price Max Age

* Settings new tests

* Fix Event on BakerfiVault

* Failed to approve allowance

* Formated output

* Stage prices

* PythOracle

* PythOracle

* Settings SDK working

* Fixed the javascript interface

* bakerFi Vault

* Vault

* Tested Stale Prices

* Working

* Chainlink Oracles

* Receive ETH only allowed by the strategy

* Funds Lost

* Fixes for comments

* Another fix

* Chainlink Oracle

+++ contracts/interfaces/core/IOracle.sol
@@ -12,7 +12,7 @@ pragma solidity ^0.8.18;
 interface IOracle {
     struct Price {
         uint256 price;
-        uint lastUpdate;
+        uint256 lastUpdate;
     }
 
     /**
